refactor : 쉐어 편집 통합 테스트 분리
